task 8

+++ project/apps/blog/src/app/blog-comment/blog-comment.service.ts
@@ -12,13 +12,17 @@ export class BlogCommentService {
     private readonly blogCommentRepository: BlogCommentRepository
   ) {}
 
-  async createComment(dto: CreateCommentDto): Promise<Comment> {
-    const commentEntity = new BlogCommentEntity(dto);
+  async createComment(dto: CreateCommentDto, userId: string): Promise<Comment> {
+    const commentEntity = new BlogCommentEntity({ ...dto, userId });
     return this.blogCommentRepository.create(commentEntity);
   }
 
-  async deleteComment(id: number): Promise<void> {
-    this.blogCommentRepository.destroy(id);
+  async deleteComment(id: number, userId: string): Promise<void> {
+    const deletingComment = await this.getComment(id);
+
+    if (deletingComment.userId === userId) {
+      this.blogCommentRepository.destroy(id);
+    }
   }
 
   async getComment(id: number): Promise<Comment> {
@@ -29,7 +33,11 @@ export class BlogCommentService {
     return this.blogCommentRepository.find(query);
   }
 
-  async updateComment(id: number, dto: UpdateCommentDto): Promise<Comment> {
-    return this.blogCommentRepository.update(id, new BlogCommentEntity(dto));
+  async updateComment(id: number, dto: UpdateCommentDto, userId: string): Promise<Comment> {
+    const editingComment = await this.getComment(id);
+
+    if (editingComment?.userId === userId) {
+      return this.blogCommentRepository.update(id, new BlogCommentEntity(dto));
+    }
   }
 }
